+++ CHANGES.md
@@ -26,6 +26,12 @@
 
 - `Cesium3DTileset.disableCollision` has been deprecated and will be removed in 1.116. Use `Cesium3DTileset.enableCollision` instead.
 
+#### @cesium/engine
+
+##### Additions :tada:
+
+- Added `I3SBuildingSceneLayerExplorer` widget for working with I3S Building Scene Layer data. [#11678](https://github.com/CesiumGS/cesium/pull/11678)
+
 ### 1.114 - 2024-02-01
 
 #### @cesium/engine